+++ Sprint 3/BackendGeems/BackendGeems/Domain/Pago.cs
@@ -6,11 +6,11 @@ namespace BackendGeems.Domain
     {
         public Guid Id { get; set; }
         public DateTime FechaRealizada { get; set; }  
-        public decimal MontoPago { get; set; }
+        public double MontoPago { get; set; }
         public Guid IdEmpleado { get; set; }
         public Guid IdPayroll { get; set; }
         public Guid IdPlanilla { get; set; }
-        public decimal MontoBruto { get; set; }
+        public double MontoBruto { get; set; }
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFinal { get; set; }
     }
